backend/requirements.txt
-i https://pypi.org/simple
aiodns==3.1.1
aiosmtpd==1.4.5; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
atpublic==4.0; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
cffi==1.16.0; python_version >= '3.8'
django==5.0.3; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
pillow==10.2.0; python_version >= '3.8'
pycares==4.4.0; python_version >= '3.8'
pycparser==2.21
sqlparse==0.4.4; python_version >= '3.5'
validate-email==1.3
verify-email==2.4.3
